Added optional type icons to type display, added expAll Redist.
ExpAll Redistribution is an optional feature to modify the expAll to grant a higher percentage of the total experience rewarded to your lower leveled pokemon. Weight values are applied and adjusted with the slider in the options menu, determining the 'strength' of this redistribution. 0 = Off, 1 = Very minimal shift, 10 = maximum shift.
Logic for expAll_redist located in '004_Battle_ExpAndMoveLearning.rb'